refactor(route): extract helper out of route_search_provider (Refs #563 phase: route_search_provider)
Move RouteSearchResult and strategyFor() factory out of the provider
file into sibling domain files. The provider re-exports both, so
existing callers and the generated .g.dart keep working unchanged.

route_search_provider.dart: 263 LOC -> 229 LOC (under the 250 cap).

Refs #563
